src/web_service/requirements.txt
Django
djangorestframework
mysqlclient
mysql-connector
